+++ core/api/hooks/index.js
@@ -1,14 +1,14 @@
-export * from './hooks.account'
-export * from './hooks.authentication'
-export * from './hooks.authorisations'
-export * from './hooks.devices'
-export * from './hooks.groups'
-export * from './hooks.logger'
-export * from './hooks.model'
-export * from './hooks.organisations'
-export * from './hooks.pusher'
-export * from './hooks.query'
-export * from './hooks.service'
-export * from './hooks.storage'
-export * from './hooks.tags'
-export * from './hooks.users'
+export * from './hooks.account.js'
+export * from './hooks.authentication.js'
+export * from './hooks.authorisations.js'
+export * from './hooks.devices.js'
+export * from './hooks.groups.js'
+export * from './hooks.logger.js'
+export * from './hooks.model.js'
+export * from './hooks.organisations.js'
+export * from './hooks.pusher.js'
+export * from './hooks.query.js'
+export * from './hooks.service.js'
+export * from './hooks.storage.js'
+export * from './hooks.tags.js'
+export * from './hooks.users.js'
