+++ core/api/models/tags.model.mongodb.js
@@ -1,4 +1,4 @@
-module.exports = function (app, options) {
+export default function (app, options) {
   const db = options.db || app.db
   options.Model = db.collection('tags')
   // Use compound index to have unique pairs scope/value
