+++ core/api/models/users.model.mongodb.js
@@ -1,4 +1,4 @@
-module.exports = function (app, options) {
+export default function (app, options) {
   options.Model = app.db.collection('users')
   options.Model.createIndex({ email: 1 }, { unique: true })
   // Collation provided in query ensure sorting to be case insensitive w.r.t. user's language
