+++ core/api/services/devices/devices.hooks.js
@@ -1,6 +1,6 @@
 import { disallow } from 'feathers-hooks-common'
 
-module.exports = {
+export default {
   before: {
     all: [],
     find: [],
