+++ core/api/services/pusher/pusher.channels.js
@@ -1,3 +1,3 @@
-module.exports = {
+export default {
   all: (data, context) => context.app.channel('authenticated').filter(connection => false)
 }
