+++ core/common/index.js
@@ -2,8 +2,8 @@
 // Tracked here : https://github.com/babel/babel/issues/2877
 // We tested the workaround given here https://github.com/babel/babel/issues/2877#issuecomment-270700000 with success so far
 import _ from 'lodash'
-import * as errors from './errors'
-import * as permissions from './permissions'
+import * as errors from './errors.js'
+import * as permissions from './permissions.js'
 
 export { errors }
 export { permissions }
