+++ core/common/permissions.js
@@ -1,5 +1,7 @@
 import _ from 'lodash'
-import { Ability, AbilityBuilder, toMongoQuery } from 'casl/dist/umd'
+import casl from 'casl/dist/umd/index.js'
+
+const { Ability, AbilityBuilder, toMongoQuery } = casl
 
 // Define some alias to simplify ability definitions
 Ability.addAlias('update', 'patch')
